Move InterpreterExpression to its own file
