+++ src/DynamicExpresso.Core/Parsing/Parser.cs
@@ -166,17 +166,6 @@ namespace DynamicExpresso.Parsing
 			}
 		}
 
-		private class ParameterWithPosition : Parameter
-		{
-			public ParameterWithPosition(int pos, string name, Type type)
-				: base(name, type)
-			{
-				Position = pos;
-			}
-
-			public int Position { get; }
-		}
-
 		private ParameterWithPosition[] ParseLambdaParameterList()
 		{
 			var hasOpenParen = _token.id == TokenId.OpenParen;
@@ -2196,7 +2185,7 @@ namespace DynamicExpresso.Parsing
 					if (method.PromotedParameters[i] is InterpreterExpression ie)
 					{
 						var actualParamInfo = actualMethodParameters[i];
-						var lambdaExpr = GenerateLambdaFromInterpreterExpression(ie, actualParamInfo.ParameterType);
+						var lambdaExpr = ie.EvalAs(actualParamInfo.ParameterType);
 						if (lambdaExpr == null)
 							return false;
 
@@ -2211,11 +2200,6 @@ namespace DynamicExpresso.Parsing
 			}
 
 			return true;
-		}
-
-		private static LambdaExpression GenerateLambdaFromInterpreterExpression(InterpreterExpression ie, Type delegateType)
-		{
-			return ie.EvalAs(delegateType);
 		}
 
 		private static MethodInfo MakeGenericMethod(MethodData method)
@@ -2313,7 +2297,7 @@ namespace DynamicExpresso.Parsing
 					return null;
 
 				if (!type.ContainsGenericParameters)
-					return GenerateLambdaFromInterpreterExpression(ie, type);
+					return ie.EvalAs(type);
 
 				return expr;
 			}
@@ -3076,77 +3060,6 @@ namespace DynamicExpresso.Parsing
 
 			var conversionType = typeof(Nullable<>).MakeGenericType(exprType);
 			return Expression.ConvertChecked(expr, conversionType);
-		}
-
-		/// <summary>
-		/// Expression that wraps over an interpreter. This is used when parsing a lambda expression 
-		/// definition, because we don't know the parameters type before resolution.
-		/// </summary>
-		private class InterpreterExpression : Expression
-		{
-			private readonly Interpreter _interpreter;
-			private readonly string _expressionText;
-			private readonly IList<Parameter> _parameters;
-			private Type _type;
-
-			public InterpreterExpression(ParserArguments parserArguments, string expressionText, params ParameterWithPosition[] parameters)
-			{
-				var settings = parserArguments.Settings.Clone();
-				_interpreter = new Interpreter(settings);
-				_expressionText = expressionText;
-				_parameters = parameters;
-
-				// Take the parent expression's parameters and set them as an identifier that
-				// can be accessed by any lower call
-				// note: this doesn't impact the initial settings, because they're cloned
-				foreach (var dp in parserArguments.DeclaredParameters)
-				{
-					// Have to mark the parameter as "Used" otherwise we can get a compilation error.
-					parserArguments.TryGetParameters(dp.Name, out var pe);
-					_interpreter.SetIdentifier(new Identifier(dp.Name, pe));
-				}
-
-				foreach (var myParameter in parameters)
-				{
-					if (settings.Identifiers.ContainsKey(myParameter.Name))
-					{
-						throw new ParseException($"A local or parameter named '{myParameter.Name}' cannot be declared in this scope because that name is used in an enclosing local scope to define a local or parameter", myParameter.Position);
-					}
-				}
-
-				// prior to evaluation, we don't know the generic arguments types
-				_type = ReflectionExtensions.GetFuncType(parameters.Length);
-			}
-
-			public IList<Parameter> Parameters
-			{
-				get { return _parameters; }
-			}
-
-			public override Type Type
-			{
-				get { return _type; }
-			}
-
-			internal LambdaExpression EvalAs(Type delegateType)
-			{
-				if (!IsCompatibleWithDelegate(delegateType))
-					return null;
-
-				var lambdaExpr = _interpreter.ParseAsExpression(delegateType, _expressionText, _parameters.Select(p => p.Name).ToArray());
-				_type = lambdaExpr.Type;
-				return lambdaExpr;
-			}
-
-			internal bool IsCompatibleWithDelegate(Type target)
-			{
-				if (!target.IsGenericType || target.BaseType != typeof(MulticastDelegate))
-					return false;
-
-				var genericTypeDefinition = target.GetGenericTypeDefinition();
-				return genericTypeDefinition == ReflectionExtensions.GetFuncType(_parameters.Count)
-					|| genericTypeDefinition == ReflectionExtensions.GetActionType(_parameters.Count);
-			}
 		}
 
 		/// <summary>
